🎨 style(server.ts): change port variable case from lowercase port to uppercase PORT ✨ feat(server.ts): add support for process.env.PORT environment variable The port variable is now named PORT, which improves consistency with the naming conventions as PORT is a constant. Support for an environment variable allows the application to be more flexible as it can now run on any available port specified via the process.env.PORT environment variable.
🐛 fix(.gitignore): ignore config.json and config-*.json files
The .gitignore file now ignores the config.json and config-*.json files.

✨ feat(app.d.ts): add model property to CarInfo interface
The CarInfo interface now includes a model property.

🎨 style(card.styled.ts): set height to 100% for CarItemWrapper
The height of CarItemWrapper is now set to 100%.

✨ feat(checkbox): add Checkbox component
The Checkbox component has been added to the project

🎨 style(cars.data.json): update cars data with new names and models
🆕 feat(color.data.json): add color data file
🎨 style(fbox.styled.ts): add gx and gy props to FBoxStyledProps
🎨 style(motorcycles/index.tsx): add absolute position to Price component
🎨 style(main-page/index.tsx): add BadgeWrapper component to wrap Badge component
The cars.data.json file has been updated with new names and models for the cars. The color.data.json file has been added to the project. The FBoxStyledProps interface has been updated to include gx and gy props. The Price component in motorcycles/index.tsx has been given an absolute position. The BadgeWrapper component has been added to main-page/index.tsx to wrap the Badge component.

🎨 style(mainpage.styled.ts): add BadgeWrapper, Badge, and Color styled components
✨ feat(special-offers/index.tsx): add ClipBox component to wrap Icon and Text components
The BadgeWrapper, Badge, and Color styled components are added to improve the styling of the main page. The BadgeWrapper is used to wrap the Badge component, which displays a badge with a number. The Color component is currently empty but will be used to add color to the page. The ClipBox component is added to wrap the Icon and Text components in the special-offers/index.tsx file. This improves the layout of the page and makes it easier to read.

🎨 style(available-cars): add color selection for standard car color
✨ feat(available-cars): add test drive and special offer checkboxes
🐛 fix(globalStyle.ts): remove extra semicolon in cursor property
The changes add a color selection for standard car color, which allows the user to choose from a list of available colors. The checkboxes for test drive and special offer are also added, which allow the user to toggle these options on and off. Additionally, a bug in globalStyle.ts is fixed by removing an extra semicolon in the cursor property.